# Conflicts:
#	src/adminTools/Modify.java
#	src/adminTools/ModifyLibrary.java
#	src/librarianTools/BookingSystem.java
#	src/librarianTools/Modify.java
#	src/librarianTools/ModifyLibrary.java
#	src/librarianTools/OutstandingRequest.java
#	src/librarianTools/ReturningSystem.java
#	src/patronTools/BookingSystem.java
#	src/patronTools/ReturningSystem.java
#	src/tools/Constants.java
#	src/tools/Logic.java
#	src/users/Librarian.java
#	src/users/Patron.java

+++ src/librarianTools/OutstandingRequest.java
@@ -2,6 +2,7 @@ package librarianTools;
 
 import documents.Document;
 import tools.Database;
+import tools.Logic;
 import tools.Request;
 
 import java.util.ArrayList;
@@ -9,41 +10,44 @@ import java.util.Date;
 import java.util.List;
 
 public class OutstandingRequest {
-	public void makeOutstandingRequest(Request request, Database database) {
+    public void makeOutstandingRequest(int librarianId, Request request, Database database) {
+        if(Logic.canSetRequest(librarianId, database)) {
+            sendNotificationsForOutstandingRequest(request, database);
+            database.deleteRequestsForDocument(request.getIdDocument());
+        } else {
 
-		sendNotificationsForOutstandingRequest(request, database);
-		database.deleteRequestsForDocument(request.getIdDocument());
+        }
 
-	}
+    }
 
-	public void setAvailability(int docID, Database database) {
-		if (database.getDocument(docID).getNumberOfCopies() > 0) {
-			sendNotificationsForAvailability(docID, database);
-		}
+    public void setAvailability(int docID, Database database)  {
+            if (database.getDocument(docID).getNumberOfCopies() > 0) {
+                sendNotificationsForAvailability(docID, database);
+            }
 
-	}
+    }
 
-	private void sendNotificationsForOutstandingRequest(Request request, Database db) {
-		ArrayList<Request> requests = null;
-		requests = db.getRequestsForDocument(request.getIdDocument());
-		for (Request temp : requests) {
-			if (temp.getIdPatron() == request.getIdPatron()) {
-				Document doc = db.getDocument(request.getIdDocument());
-				db.insertNotification(temp.getRequestId(), temp.getIdPatron(),
-						"Outstanding request for " + doc.getTitle(), new Date());
-			}
-		}
+    private void sendNotificationsForOutstandingRequest(Request request, Database db) {
+        ArrayList<Request> requests = null;
+            requests = db.getRequestsForDocument(request.getIdDocument());
+            for (Request temp : requests) {
+                if (temp.getIdPatron() == request.getIdPatron()) {
+                    Document doc = db.getDocument(request.getIdDocument());
+                    db.insertNotification(temp.getRequestId(), temp.getIdPatron(),
+                            "Outstanding request for " + doc.getTitle(), new Date());
+                }
+            }
 
-	}
+    }
 
-	private void sendNotificationsForAvailability(int docId, Database database) {
-		List<Request> requests = null;
-		requests = database.getRequestsForDocument(docId);
-		int i = 0;
-		while (i < database.getDocument(docId).getNumberOfCopies()) {
-			Request temp = requests.get(i);
-			database.insertNotification(temp.getRequestId(), temp.getIdPatron(), "Set available document", new Date());
-			i++;
-		}
-	}
+    private void sendNotificationsForAvailability(int docId, Database database) {
+        List<Request> requests = null;
+            requests = database.getRequestsForDocument(docId);
+            int i = 0;
+            while (i < database.getDocument(docId).getNumberOfCopies()) {
+                Request temp = requests.get(i);
+                database.insertNotification(temp.getRequestId(), temp.getIdPatron(), "Set available document", new Date());
+                i++;
+            }
+    }
 }
